fix(commands/commit): check cz commit msg
